# String: don't materialize Regex match[0] if not needed