This avoid creating an intermediate string that's produced by `MatchData#[]` that's completely not needed for the task.

Here's a benchmark:

```crystal
require "benchmark"

Benchmark.ips do |x|
  x.report("reverse") do
    "há日本語".reverse
  end
  x.report("scan") do
    "foo".scan(/o+/) { }
  end
  x.report("split") do
    "foboloxoro".split(/o/) { }
  end
end
```

Results before this PR:

```
reverse   1.29M (773.61ns) (± 5.74%)   384B/op 
   scan   6.53M (153.11ns) (± 7.77%)  48.0B/op
  split   1.59M (628.53ns) (± 5.13%)   241B/op 
```

Results after this PR:

```
reverse   2.23M (447.64ns) (± 3.66%)   128B/op
   scan   8.76M (114.15ns) (± 4.63%)  32.0B/op
  split   2.22M (450.89ns) (± 2.36%)   160B/op
```

Note in all these cases less memory is used now, and it's faster.